moved lib_loc_info tm generator
