+++ umich_catalog_indexing/lib/jobs/translation_map_generator.rb
@@ -1,4 +1,5 @@
 require_relative "translation_map_generator/electronic_collections"
+require_relative "translation_map_generator/lib_loc_info"
 module Jobs
   module TranslationMapGenerator
   end
