+++ umich_catalog_indexing/spec/jobs/translation_map_generator/libLocInfo_spec.rb
@@ -1,15 +1,15 @@
-require 'yaml'
-require_relative '../spec_helper.rb'
-require 'jobs'
-RSpec.describe Jobs::LibLocInfo::LibraryLocationList, "#list" do
+require "yaml"
+require_relative "../../spec_helper"
+require "jobs"
+RSpec.describe Jobs::TranslationMapGenerator::LibLocInfo::LibraryLocationList, "#list" do
   before(:each) do
-    @libraries = JSON.parse(fixture('/aael.json'))
-    @locations = JSON.parse(fixture('/aael_locations_short.json'))
-    @output = YAML.load_file('spec/fixtures/libLocShort.yaml')
+    @libraries = JSON.parse(fixture("/aael.json"))
+    @locations = JSON.parse(fixture("/aael_locations_short.json"))
+    @output = YAML.load_file("spec/fixtures/libLocShort.yaml")
   end
   subject do
-    stub_alma_get_request(url: 'conf/libraries', output: @libraries.to_json)
-    stub_alma_get_request(url: 'conf/libraries/AAEL/locations', output: @locations.to_json)
+    stub_alma_get_request(url: "conf/libraries", output: @libraries.to_json)
+    stub_alma_get_request(url: "conf/libraries/AAEL/locations", output: @locations.to_json)
     described_class.new.list
   end
   it "returns expected yaml" do
@@ -43,14 +43,14 @@ RSpec.describe Jobs::LibLocInfo::LibraryLocationList, "#list" do
     @output["AAEL DISP"]["name"] = "Art Architecture & Engineering"
     expect(subject).to eq(@output)
   end
-  it "doesn't print location code if it's the internal name and not upcased" do 
+  it "doesn't print location code if it's the internal name and not upcased" do
     @locations["location"][0]["external_name"] = ""
     @locations["location"][0]["name"] = "DiSp"
     @output["AAEL DISP"]["name"] = "Art Architecture & Engineering"
     expect(subject).to eq(@output)
   end
   it "skips numerical locations" do
-    @locations["location"][0]["code"] = '9999'
+    @locations["location"][0]["code"] = "9999"
     expect(subject.keys.count).to eq(1)
   end
 end
